requirements.txt
blas==1.0
brotli-python==1.0.9
ca-certificates==2023.12.12
certifi==2024.2.2
charset-normalizer==2.0.4
colorama==0.4.6
cudatoolkit==11.3.1
faiss-cpu==1.7.4
faiss==1.7.4
freetype==2.12.1
idna==3.4
intel-openmp==2023.1.0
jinja2==3.1.2
joblib==1.3.2
jpeg==9e
lerc==3.0
libdeflate==1.17
libfaiss==1.7.4
libpng==1.6.39
libtiff==4.5.1
libuv==1.44.2
libwebp-base==1.3.2
lz4-c==1.9.4
markupsafe==2.1.3
mkl-service==2.4.0
mkl==2023.1.0
mkl_fft==1.3.8
mkl_random==1.2.4
numpy-base==1.24.3
numpy==1.24.4
openjpeg==2.4.0
openssl==1.1.1w
pillow==10.2.0
pip==23.3.1
pyparsing==3.1.1
pysocks==1.7.1
python==3.8.5
pytorch-mutex==1.0
pytorch==1.11.0
pyyaml==6.0.1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.10.1
setuptools==68.2.2
sqlite==3.41.2
tbb==2021.8.0
threadpoolctl==3.2.0
torch-cluster==1.6.0
torch-geometric==2.1.0
torch-scatter==2.0.9
torch-sparse==0.6.15
torch-spline-conv==1.2.1
torchaudio==0.11.0
torchsummary==1.5.1
torchvision==0.12.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
vc==14.2
vs2015_runtime==14.27.29016
wheel==0.41.2
win_inet_pton==1.1.0
xz==5.4.6
yaml==0.2.5
zlib==1.2.13
zstd==1.5.5

torchdiffeq~=0.2.4